das-project/requirements.txt
# === Core ===
streamlit>=1.40.0
supabase>=2.0.0
python-dotenv>=1.0.0

# === Data ===
pandas>=2.0.0
openpyxl>=3.1.0

# === Test Data Generation ===
faker>=30.0.0

# === Testing ===
pytest>=8.0.0
pytest-cov>=4.0.0

# === Utilities ===
python-dateutil>=2.8.0
